Update GraphQL packages and implement GraphQL client
Added additional dependencies such as 'graphql' and 'urql' to the package.json and updated yarn.lock accordingly. Created a new file 'api-provider.tsx' to establish a URQL client for future GraphQL API calls. The client is then used in the application layout, wrapped around the existing provider structure. This change facilitates communication with a GraphQL server for fetching and updating data in the application.