Refactor graphql codegen output path
Refactored the output path for graphql codegen from 'src/gql/' to 'src/graphql-generated/'. As a result, 'src/gql/graphql.ts' was deleted and replaced with 'src/graphql-generated/graphql.ts'. All references were updated accordingly to maintain consistency across the project. This change was made to abide by a naming convention that better reflects the contents and purpose of the directory.