Add mapService utility and refactor sign-up form error handling
Introduced a MapService utility class in the utils folder for transforming GraphQL FieldErrors into a form-friendly object. This allows for a more streamlined error handling in forms. Specifically, this was implemented in the sign-up page. Here, FieldErrors returned from the sign-up mutation are now converted to form errors using the MapService. Additionally, unhandled server errors are now handled as global errors, which are presented as alerts to the user.